# docs/swagger.yaml
basePath: /smrp
definitions:
...
  dto.ReportQueryDto:
    properties:
      _limit:
        type: integer
      _page:
        type: integer
      datefrom:
        type: string
      dateto:
        type: string
      vt:
        type: integer
    required:
    - datefrom
    - dateto
    type: object
...
info:
  contact: {}
  description: SMRP Backend API description.
  title: Swagger SMRP Backend API
  version: "1.0"
paths:
...
  /api/master-pd105/export/rpt1:
    get:
      parameters:
      - description: datefrom
        in: query
        name: datefrom
        type: string
      - description: dateto
        in: query
        name: dateto
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Report/MasterPD105
  /api/master-pd105/export/rpt1/xlsx:
    get:
      parameters:
      - description: datefrom
        in: query
        name: datefrom
        type: string
      - description: dateto
        in: query
        name: dateto
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Report/MasterPD105
  /api/master-pd105/rpt1:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      - description: datefrom
        in: query
        name: datefrom
        type: string
      - description: dateto
        in: query
        name: dateto
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Report/MasterPD105
    post:
      parameters:
      - description: Search Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReportQueryDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Report/MasterPD105
  /api/master-pd105/rpt1/{id}:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Report/MasterPD105
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Report/MasterPD105
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"